requirements.txt
runpod
transformers
accelerate
pillow
timm
einops
sentencepiece
protobuf
safetensors
